% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anycomp.R
\name{anycomp}
\alias{anycomp}
\title{Scrapes Competition Event Data from USFSA website}
\usage{
anycomp(data)
}
\arguments{
\item{data}{html web link, passed to \code{\link{read_html}}}
}
\value{
data frame of usable Event data
}
\description{
Scrapes Competition Event Data from USFSA website
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
}
