% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bothcomp.R
\name{bothcomp}
\alias{bothcomp}
\title{Skaters who Participated in Both Events}
\usage{
bothcomp(dat1, dat2)
}
\arguments{
\item{dat1}{data frame of Event 1}

\item{dat2}{data frame of Event 2}
}
\value{
data frame of skaters who participated in both chosen events
}
\description{
Skaters who Participated in Both Events
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
Comp2 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2019/27958/SEGM040.html')
bothcomp(Comp1,Comp2)
}
