% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvcompareComp.R
\name{bvcompareComp}
\alias{bvcompareComp}
\title{Comparing Base Value Scores for two Events}
\usage{
bvcompareComp(dat1, dat2)
}
\arguments{
\item{dat1}{data frame of first event}

\item{dat2}{data frame of second event}
}
\value{
boxplots for two competitions
}
\description{
Comparing Base Value Scores for two Events
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
Comp2 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2019/27958/SEGM040.html')
bvcompareComp(Comp1, Comp2)
}
