% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvcompareSkater.R
\name{bvcompareSkater}
\alias{bvcompareSkater}
\title{Histogram of Base Value Scores by Skater}
\usage{
bvcompareSkater(data)
}
\arguments{
\item{data}{data frame of chosen event data}
}
\value{
histogram of Base Value Scores for a single competition
}
\description{
Histogram of Base Value Scores by Skater
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
bvcompareSkater(Comp1)
}
