% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compsumQuan.R
\name{compsumQuan}
\alias{compsumQuan}
\title{Calculates Mean & Standard Deviation for Quantitative Variables for One Event}
\usage{
compsumQuan(dat)
}
\arguments{
\item{dat}{data frame of event data}
}
\value{
data frame of summary statistics of quantitative variables for a single event
}
\description{
Calculates Mean & Standard Deviation for Quantitative Variables for One Event
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
compsumQuan(Comp1)
}
