% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goecompare.R
\name{goecompare}
\alias{goecompare}
\title{Histogram of GOE Scores by Skater}
\usage{
goecompare(data)
}
\arguments{
\item{data}{data frame of chosen event data}
}
\value{
histogram of GOE Scores for a single competition
}
\description{
Histogram of GOE Scores by Skater
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
goecompare(Comp1)
}
