% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runEventApp}
\alias{runEventApp}
\title{Run Shiny App with preset USFSA Events to choose from}
\usage{
runEventApp()
}
\value{
shiny application
}
\description{
Run Shiny App with preset USFSA Events to choose from
}
