% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsscompareComp.R
\name{tsscompareComp}
\alias{tsscompareComp}
\title{Comparing Total Segment Scores for two Events}
\usage{
tsscompareComp(dat1, dat2)
}
\arguments{
\item{dat1}{data frame of first event data}

\item{dat2}{data frame of second event data}
}
\value{
boxplot for two competitions
}
\description{
Comparing Total Segment Scores for two Events
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
Comp2 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2019/27958/SEGM040.html')
tsscompareComp(Comp1, Comp2)
}
