% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsscompareSkater.R
\name{tsscompareSkater}
\alias{tsscompareSkater}
\title{Histogram of Total Segment Scores by Skater}
\usage{
tsscompareSkater(data)
}
\arguments{
\item{data}{data frame of chosen event data}
}
\value{
histogram of Total Segment Scores for a single competition
}
\description{
Histogram of Total Segment Scores by Skater
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
tsscompareSkater(Comp1)
}
