% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varvariationQual.R
\name{varvariationQual}
\alias{varvariationQual}
\title{Bar graph for Qualitive Variable: Deductions in one Event}
\usage{
varvariationQual(data)
}
\arguments{
\item{data}{data frame of event data}
}
\value{
bar graph for the qualitative variable: Deductions for a single event
}
\description{
Bar graph for Qualitive Variable: Deductions in one Event
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
varvariationQual(Comp1)
}
