% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varvariationQuan.R
\name{varvariationQuan}
\alias{varvariationQuan}
\title{Boxplots for Quantitive Variables for one Event}
\usage{
varvariationQuan(data)
}
\arguments{
\item{data}{data frame of event data}
}
\value{
boxplot of quantitative variables for a single event
}
\description{
Boxplots for Quantitive Variables for one Event
}
\examples{
Comp1 = anycomp('https://ijs.usfigureskating.org/leaderboard/results/2018/26192/SEGM036.html')
varvariationQuan(Comp1)
}
